backend/requirements.txt
APScheduler==3.11.0
beautifulsoup4==4.14.3
chromadb==1.3.7
fastapi==0.127.0
lxml==6.0.2
openai==2.14.0
playwright==1.55.0
pydantic==2.12.5
python-dotenv==1.2.1
